fixed location tests, added comment tests

+++ poi-basicAuth/app/api/locations.js
@@ -42,6 +42,7 @@ const Locations = {
       strategy: 'jwt',
     },
     handler: async function(request, h) {
+      console.log("create location API handler at work");
       const userId = utils.getUserIdFromRequest(request);
       const newLocation = new Location(request.payload);
       newLocation.author = userId;
@@ -71,14 +72,24 @@ const Locations = {
       strategy: 'jwt',
     },
     handler: async function(request, h) {
-      //const response = await Location.remove({ _id: request.params.id });
-      const response = await Location.deleteOne({ _id: request.params.id });
-      if (response.deletedCount === 1) {
-        return { success: true };
+      console.log("delete location handler at work");
+      const userId = utils.getUserIdFromRequest(request);
+      console.log("user id is: " + userId);
+      const returnedLocation = await Location.findOne({ _id: request.params.id });
+      console.log("location author id is: " + returnedLocation.author);
+      if (userId === returnedLocation.author) {
+        const response = await Location.deleteOne({ _id: request.params.id });
+        if (response.deletedCount === 1) {
+          console.log("delete was successful");
+          return { success: true };
+        }
+        console.log("userId and author id don't match");
+        return Boom.badRequest('userId not a match for author id');
       }
       return Boom.notFound('id not found');
     }
-  }
+  },
+
 };
 
 module.exports = Locations;
